+++ src/components/data-entry/QueryItem/Text.tsx
@@ -3,6 +3,10 @@ import { Typography } from 'src/components/general/Typography/Typography'
 export interface ITextProps {
   disabled?: boolean
   text: string
+  /**
+   * @deprecated Not being used by the component, included here to fix runtime typing problems. Needs to be worked on
+   */
+  className?: string
 }
 
 const Text = ({ disabled = false, text }: ITextProps) => {