+++ src/types/icons.ts
@@ -1,13 +1,18 @@
 import type { RequireAtLeastOne } from 'type-fest'
 
-export type IconVariant = 'light' | 'duo-tone'
+export const ICON_VARIANTS = {
+  LIGHT: 'light',
+  DUO_TONE: 'duo-tone',
+} as const
+
+export type IconVariant = (typeof ICON_VARIANTS)[keyof typeof ICON_VARIANTS]
 
 export type IconOptions = RequireAtLeastOne<
   {
-    light?: React.ComponentType<React.SVGProps<SVGSVGElement>>
-    'duo-tone'?: React.ComponentType<React.SVGProps<SVGSVGElement>>
+    [ICON_VARIANTS.LIGHT]?: React.ComponentType<React.SVGProps<SVGSVGElement>>
+    [ICON_VARIANTS.DUO_TONE]?: React.ComponentType<React.SVGProps<SVGSVGElement>>
     default: IconVariant
-    deprecated?: string
+    deprecated?: string | boolean
   },
   IconVariant
 >
@@ -16,13 +21,16 @@ export type IconOptions = RequireAtLeastOne<
 
 export type IconNames =
   | 'abSplit'
+  | 'accept'
   | 'account'
   | 'active'
   | 'agentCopilot'
   | 'add'
   | 'alicorn'
+  | 'audienceGroup'
   | 'analytics'
   | 'annotation'
+  | 'array'
   | 'api'
   | 'audiences'
   | 'C360'
@@ -37,6 +45,7 @@ export type IconNames =
   | 'cohort'
   | 'connections'
   | 'conversion'
+  | 'criteriaGroup'
   | 'copy'
   | 'dashboard'
   | 'database'
@@ -69,6 +78,7 @@ export type IconNames =
   | 'heart'
   | 'help'
   | 'helpVideo'
+  | 'history'
   | 'identity'
   | 'info'
   | 'insights'
@@ -77,6 +87,7 @@ export type IconNames =
   | 'jumpTo'
   | 'lightBulb'
   | 'link'
+  | 'list'
   | 'liveStream'
   | 'lock'
   | 'logout'
@@ -86,16 +97,21 @@ export type IconNames =
   | 'myHub'
   | 'next'
   | 'notification'
+  | 'number'
   | 'notificationSubscribed'
   | 'notificationSubscribe'
   | 'observability'
   | 'openTab'
   | 'organization'
+  | 'other'
+  | 'otherData'
   | 'oversight'
   | 'overview'
   | 'paywall'
+  | 'pause'
   | 'pipelines'
   | 'placeholder'
+  | 'play'
   | 'precision'
   | 'predictions'
   | 'premium'
@@ -107,6 +123,7 @@ export type IconNames =
   | 'rateDown'
   | 'rateStar'
   | 'rateUp'
+  | 'run'
   | 'remove'
   | 'savedProjects'
   | 'scheduledReport'
@@ -121,20 +138,25 @@ export type IconNames =
   | 'siteMap'
   | 'sparkles'
   | 'split'
+  | 'string'
   | 'support'
   | 'systemAlerts'
   | 'textWidget'
+  | 'timestamp'
   | 'transformation'
   | 'trends'
   | 'unlock'
   | 'upload'
+  | 'user'
   | 'userAttribute'
   | 'userProfiles'
   | 'users'
+  | 'viewOnly'
   | 'wrench'
   | 'zoomIn'
   | 'zoomOut'
   | 'bannerFreemium'
+  | 'boolean'
   | 'stateEmpty'
   | 'stateError'
   | 'stateNoResults'
@@ -144,3 +166,4 @@ export type IconNames =
   | 'clone'
   | 'reorder'
   | 'refresh'
+  | 'moveToTop'